+++ R/bootstrap_model.R
@@ -31,14 +31,14 @@
 #' slopes, etc. and their confidence intervals. These can then be passed to
 #' `model_parameter()` to obtain standard errors, p-values, etc. (see
 #' example).
-#' \cr\cr
+#'
 #' Note that that p-values returned here are estimated under the assumption of
 #' *translation equivariance*: that shape of the sampling distribution is
 #' unaffected by the null being true or not. If this assumption does not hold,
 #' p-values can be biased, and it is suggested to use proper permutation tests
 #' to obtain non-parametric p-values.
 #'
-#' @seealso [bootstrap_parameters()], [simulate_model()], [simulate_parameters()]
+#' @seealso [`bootstrap_parameters()`], [`simulate_model()`], [`simulate_parameters()`]
 #'
 #' @examples
 #' \dontrun{