+++ R/methods_ordinal.R
@@ -130,7 +130,7 @@ simulate_model.clm2 <- function(model,
                                 component = c("all", "conditional", "scale"),
                                 ...) {
   component <- match.arg(component)
-  out <- .simulate_model(model, iterations, component = component)
+  out <- .simulate_model(model, iterations, component = component, ...)
 
   class(out) <- c("parameters_simulate_model", class(out))
   attr(out, "object_name") <- insight::safe_deparse(substitute(model))