+++ R/methods_selection.R
@@ -81,7 +81,7 @@ standard_error.selection <- function(model, component = c("all", "selection", "o
 #' @export
 simulate_model.selection <- function(model, iterations = 1000, component = c("all", "selection", "outcome", "auxiliary"), ...) {
   component <- match.arg(component)
-  out <- .simulate_model(model, iterations, component = component, effects = "fixed")
+  out <- .simulate_model(model, iterations, component = component, effects = "fixed", ...)
 
   class(out) <- c("parameters_simulate_model", class(out))
   attr(out, "object_name") <- insight::safe_deparse(substitute(model))